[#65] 리팩토링 : (breadReview) context -> jotai state 로 변경
